requirements.txt
fastapi
uvicorn
streamlit
tensorflow
numpy
pillow
pickle5
openai
pydantic
requests
